User Image Upload is done

+++ src/Tahseen.Service/Services/Users/UserService.cs
@@ -4,6 +4,7 @@ using Tahseen.Data.IRepositories;
 using Tahseen.Data.Repositories;
 using Tahseen.Domain.Entities;
 using Tahseen.Service.DTOs.Feedbacks.UserRatings;
+using Tahseen.Service.DTOs.FileUpload;
 using Tahseen.Service.DTOs.Users.BorrowedBookCart;
 using Tahseen.Service.DTOs.Users.Fine;
 using Tahseen.Service.DTOs.Users.User;
@@ -12,6 +13,7 @@ using Tahseen.Service.DTOs.Users.UserProgressTracking;
 using Tahseen.Service.DTOs.Users.UserSettings;
 using Tahseen.Service.Exceptions;
 using Tahseen.Service.Interfaces.IFeedbackService;
+using Tahseen.Service.Interfaces.IFileUploadService;
 using Tahseen.Service.Interfaces.IUsersService;
 
 namespace Tahseen.Service.Services.Users
@@ -26,10 +28,17 @@ namespace Tahseen.Service.Services.Users
         private readonly IUserRatingService _userRatingService;
         private readonly IUserCartService _userCartService;
         private readonly IBorrowBookCartService _borrowBookCartService;
+        private readonly IFileUploadService _fileUploadService;
 
-        public UserService(IRepository<User> userRepository, IMapper mapper, IFineService fineService,
-            IUserProgressTrackingService userProgressTrackingService, IUserSettingService userSettingService,
-            IUserRatingService userRatingService, IUserCartService userCartService, IBorrowBookCartService borrowBookCartService)
+        public UserService(IRepository<User> userRepository,
+            IMapper mapper,
+            IFineService fineService,
+            IUserProgressTrackingService userProgressTrackingService, 
+            IUserSettingService userSettingService,
+            IUserRatingService userRatingService, 
+            IUserCartService userCartService,
+            IBorrowBookCartService borrowBookCartService,
+            IFileUploadService fileUploadService)
         {
             this._userRepository = userRepository;
             this._mapper = mapper;
@@ -39,6 +48,7 @@ namespace Tahseen.Service.Services.Users
             this._userRatingService = userRatingService;
             this._userCartService = userCartService;
             this._borrowBookCartService = borrowBookCartService;
+            this._fileUploadService = fileUploadService;
         }
         public async Task<UserForResultDto> AddAsync(UserForCreationDto dto)
         {
@@ -47,9 +57,17 @@ namespace Tahseen.Service.Services.Users
             {
                 throw new TahseenException(400, "User is exist");
             }
-            var data = this._mapper.Map<User>(dto);
-            var CreatedData = await this._userRepository.CreateAsync(data);
 
+            var FileUloadForCreation = new FileUploadForCreationDto
+            {
+                FolderPath = "UsersAssets",
+                FormFile = dto.UserImage
+            };
+            var FileResult = await this._fileUploadService.FileUploadAsync(FileUloadForCreation);
+
+            var data = this._mapper.Map<User>(dto);
+            data.UserImage = Path.Combine("Assets", $"{FileResult.FolderPath}", FileResult.FileName);
+            var CreatedData = await this._userRepository.CreateAsync(data);
                
             var UserRatingForCreation = new UserRatingForCreationDto()
             {
@@ -59,7 +77,6 @@ namespace Tahseen.Service.Services.Users
             };
 
             await _userRatingService.AddAsync(UserRatingForCreation);
-
 
             var UserCartCreation = new UserCartForCreationDto()
             {
@@ -73,6 +90,8 @@ namespace Tahseen.Service.Services.Users
             { 
                 UserId = CreatedData.Id 
             };
+
+
             await _borrowBookCartService.AddAsync(borrowBookCartCreationDto);
 
             var UserSettingCreation = new UserSettingsForCreationDto()
@@ -93,9 +112,22 @@ namespace Tahseen.Service.Services.Users
             var data = await _userRepository.SelectAll().Where(e => e.Id == Id && e.IsDeleted == false).FirstOrDefaultAsync();
             if (data is not null)
             {
+
+                // deleting image
+                await this._fileUploadService.FileDeleteAsync(data.UserImage);
+
+                // uploading image
+                var FileUloadForCreation = new FileUploadForCreationDto
+                {
+                    FolderPath = "UsersAssets",
+                    FormFile = dto.UserImage
+                };
+                var FileResult = await this._fileUploadService.FileUploadAsync(FileUloadForCreation);
+
                 var MappedData = this._mapper.Map(dto, data);
+                MappedData.UserImage = Path.Combine("Assets", $"{FileResult.FolderPath}", FileResult.FileName);
                 MappedData.UpdatedAt = DateTime.UtcNow;
-                var result = await _userRepository.CreateAsync(MappedData);
+                var result = await _userRepository.UpdateAsync(MappedData);
                 return _mapper.Map<UserForResultDto>(result);
             }
             throw new TahseenException(404, "User is not found");
@@ -103,13 +135,25 @@ namespace Tahseen.Service.Services.Users
 
         public async Task<bool> RemoveAsync(long Id)
         {
+            var user = await this._userRepository.SelectAll()
+                .Where(u => u.Id == Id && u.IsDeleted == false)
+                .FirstOrDefaultAsync();
+
+            if (user is null)
+                throw new TahseenException(404, "User is not found");
+
+            await this._fileUploadService.FileDeleteAsync(user.UserImage);
             return await _userRepository.DeleteAsync(Id);
         }
 
         public async Task<IEnumerable<UserForResultDto>> RetrieveAllAsync()
         {
-            var AllData =  _userRepository.SelectAll().Where(t => t.IsDeleted == false);
-            return _mapper.Map<IEnumerable<UserForResultDto>>(AllData);
+            var users =  _userRepository.SelectAll().Where(t => t.IsDeleted == false);
+            foreach (var user in users)
+            {
+                user.UserImage = $"https://localhost:7020/{user.UserImage.Replace('\\', '/').TrimStart('/')}";
+            }
+            return _mapper.Map<IEnumerable<UserForResultDto>>(users);
         }
 
         public async Task<UserForResultDto> RetrieveByIdAsync(long Id)
@@ -117,6 +161,7 @@ namespace Tahseen.Service.Services.Users
             var data = await _userRepository.SelectByIdAsync(Id);
             if(data != null && data.IsDeleted == false)
             {
+                data.UserImage = $"https://localhost:7020/{data.UserImage.Replace('\\', '/').TrimStart('/')}";
                 return this._mapper.Map<UserForResultDto>(data);
             }
             throw new TahseenException(404, "User is not found");
